Update task 5 #3

+++ 0x0B-malloc_free/100-argstostr.c
@@ -18,8 +18,7 @@ char *argstostr(int ac, char **av)
 	unsigned long k;
 	unsigned long total_len = 0;
 
-	if (ac == 0 || av == NULL)
-		return (NULL);
+	if (ac == 0 || av == NULL) return (NULL);
 
 	/* find total length to allocate to new string */
 	/* by summing the lengths of each argument */
@@ -43,7 +42,7 @@ char *argstostr(int ac, char **av)
 			k++;
 		}
 
-		if (i != ac - 1) /* if we are not on the last arg */
+		if (i != (unsigned long) ac - 1) /* if we are not on the last arg */
 		{
 			string[k] = '\n'; /* add newline after arg */
 			k++;
